Move translations folder
